# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spotlight Puerto Rico Business Directory API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  business_owner: "business_owner",
  admin: "admin",
} as const;

export interface User {
  id: string;
  username?: string;
  firstName?: string;
  lastName?: string;
  profileImage?: string;
  role: UserRole;
  createdAt: string;
}

export interface GetCurrentAuthUserResponse {
  isAuthenticated: boolean;
  user?: User;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon?: string;
  businessCount?: number;
}

export interface CategoryListResponse {
  categories: Category[];
}

export type BusinessStatus =
  (typeof BusinessStatus)[keyof typeof BusinessStatus];

export const BusinessStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Business {
  id: number;
  name: string;
  slug: string;
  description?: string;
  categoryId?: number;
  categoryName?: string;
  municipality?: string;
  address?: string;
  phone?: string;
  email?: string;
  website?: string;
  logoUrl?: string;
  coverUrl?: string;
  status: BusinessStatus;
  featured: boolean;
  averageRating?: number;
  reviewCount?: number;
  ownerId?: string;
  ownerName?: string;
  createdAt: string;
  updatedAt: string;
}

export type BusinessDetailHours = { [key: string]: string };

export type BusinessDetailSocialLinks = {
  facebook?: string;
  instagram?: string;
  twitter?: string;
};

export type BusinessDetail = Business & {
  hours?: BusinessDetailHours;
  socialLinks?: BusinessDetailSocialLinks;
};

export interface BusinessListResponse {
  businesses: Business[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface AdminBusinessListResponse {
  businesses: Business[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type CreateBusinessBodyHours = { [key: string]: string };

export type CreateBusinessBodySocialLinks = {
  facebook?: string;
  instagram?: string;
  twitter?: string;
};

export interface CreateBusinessBody {
  name: string;
  description: string;
  categoryId: number;
  municipality: string;
  address?: string;
  phone?: string;
  email?: string;
  website?: string;
  logoUrl?: string;
  coverUrl?: string;
  hours?: CreateBusinessBodyHours;
  socialLinks?: CreateBusinessBodySocialLinks;
}

export type UpdateBusinessBodyHours = { [key: string]: string };

export type UpdateBusinessBodySocialLinks = {
  facebook?: string;
  instagram?: string;
  twitter?: string;
};

export interface UpdateBusinessBody {
  name?: string;
  description?: string;
  categoryId?: number;
  municipality?: string;
  address?: string;
  phone?: string;
  email?: string;
  website?: string;
  logoUrl?: string;
  coverUrl?: string;
  hours?: UpdateBusinessBodyHours;
  socialLinks?: UpdateBusinessBodySocialLinks;
}

export interface Review {
  id: number;
  businessId: number;
  userId: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  title?: string;
  body?: string;
  authorName?: string;
  authorImage?: string;
  createdAt: string;
}

export interface ReviewListResponse {
  reviews: Review[];
  total: number;
}

export interface CreateReviewBody {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  title?: string;
  body?: string;
}

export interface AdminStats {
  totalBusinesses: number;
  pendingBusinesses: number;
  approvedBusinesses: number;
  totalReviews: number;
  totalUsers: number;
}

export interface UserListResponse {
  users: User[];
  total: number;
}

export type UpdateUserRoleBodyRole =
  (typeof UpdateUserRoleBodyRole)[keyof typeof UpdateUserRoleBodyRole];

export const UpdateUserRoleBodyRole = {
  user: "user",
  business_owner: "business_owner",
  admin: "admin",
} as const;

export interface UpdateUserRoleBody {
  role: UpdateUserRoleBodyRole;
}

export type ListBusinessesParams = {
  search?: string;
  category?: string;
  municipality?: string;
  page?: number;
  limit?: number;
  featured?: boolean;
};

export type AdminListBusinessesParams = {
  status?: AdminListBusinessesStatus;
  page?: number;
  limit?: number;
};

export type AdminListBusinessesStatus =
  (typeof AdminListBusinessesStatus)[keyof typeof AdminListBusinessesStatus];

export const AdminListBusinessesStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  all: "all",
} as const;

export type AdminListReviewsParams = {
  page?: number;
  limit?: number;
};

export type AdminListUsersParams = {
  page?: number;
  limit?: number;
};