# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Spotlight Puerto Rico Business Directory API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/businesses:
    get:
      operationId: adminListBusinesses
      tags: [admin]
      summary: Admin - list all businesses
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, all]
            default: all
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: All businesses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBusinessListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/businesses/{id}/approve:
    post:
      operationId: approveBusiness
      tags: [admin]
      summary: Admin - approve a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/businesses/{id}/reject:
    post:
      operationId: rejectBusiness
      tags: [admin]
      summary: Admin - reject a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/businesses/{id}/feature:
    post:
      operationId: featureBusiness
      tags: [admin]
      summary: Admin - toggle featured status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Featured toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/reviews:
    get:
      operationId: adminListReviews
      tags: [admin]
      summary: Admin - list all reviews
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: All reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/reviews/{id}:
    delete:
      operationId: adminDeleteReview
      tags: [admin]
      summary: Admin - delete a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Admin - get dashboard stats
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin - list all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}/role:
    put:
      operationId: updateUserRole
      tags: [admin]
      summary: Admin - update user role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleBody"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        profileImage:
          type: string
        role:
          type: string
          enum: [user, business_owner, admin]
        createdAt:
          type: string
          format: date-time
      required: [id, role, createdAt]
...
    Business:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        municipality:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        logoUrl:
          type: string
        coverUrl:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        featured:
          type: boolean
        averageRating:
          type: number
        reviewCount:
          type: integer
        ownerId:
          type: string
        ownerName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, slug, status, featured, createdAt, updatedAt]
...
    AdminBusinessListResponse:
      type: object
      properties:
        businesses:
          type: array
          items:
            $ref: "#/components/schemas/Business"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [businesses, total, page, limit, totalPages]
...
    Review:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        userId:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
        body:
          type: string
        authorName:
          type: string
        authorImage:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, businessId, userId, rating, createdAt]
    ReviewListResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
      required: [reviews, total]
...
    AdminStats:
      type: object
      properties:
        totalBusinesses:
          type: integer
        pendingBusinesses:
          type: integer
        approvedBusinesses:
          type: integer
        totalReviews:
          type: integer
        totalUsers:
          type: integer
      required: [totalBusinesses, pendingBusinesses, approvedBusinesses, totalReviews, totalUsers]
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
      required: [users, total]
    UpdateUserRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [user, business_owner, admin]
      required: [role]