# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spotlight Puerto Rico Business Directory API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: string;
  username?: string;
  firstName?: string;
  lastName?: string;
  profileImage?: string;
  role: UserRole;
  createdAt: Date;
}